+++ ecs/table.go
@@ -60,7 +60,7 @@ func (t *table) Add(entity Entity) uint32 {
 	return idx
 }
 
-func (t *table) Get(component ID, index uint32) unsafe.Pointer {
+func (t *table) Get(component ID, index uintptr) unsafe.Pointer {
 	return t.columns[t.components[component.id]].Get(index)
 }
 
@@ -68,7 +68,7 @@ func (t *table) Has(component ID) bool {
 	return t.components[component.id] >= 0
 }
 
-func (t *table) GetEntity(index uint32) Entity {
+func (t *table) GetEntity(index uintptr) Entity {
 	return *(*Entity)(t.entities.Get(index))
 }
